requirements.txt
fastapi
uvicorn[standard]
scikit-learn
spacy
thefuzz
pydantic
sentence-transformers
torch
numpy
playwright
python-dotenv